requirements.txt
annotated-types==0.7.0
anyio==4.10.0
autopep8==2.3.2
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
dnspython==2.7.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mypy_extensions==1.1.0
numpy==2.3.1
openai==1.100.2
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
